// src/app/models/movimentacoes_model.ts
export default interface MovimentacoesModel {
  MOV_CODIGO: number;
  MOV_CREDITO: number;
  MOV_DEBITO: number;
  MOV_DESCRICAO: string;
  MOV_TIPO: number;
  MOV_DATA: string;
  MOV_SALDOANT: number;
  MOV_CON: number;
  MOV_DATAHORA: string;
  MOV_ORDENA: number;
  MOV_PLANO: string;
  MOV_NOME: string;
  MOV_CAI: number;
  MOV_ESTADO: string;
  MOV_TROCO: number;
  PDV: number;
}